requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
razorpay==1.3.0
